// InvoiceManagement.API/ClientApp/src/utils/api.ts
export class CreateInvoiceCommand implements ICreateInvoiceCommand {
    invoiceNumber?: string | undefined;
    from?: string | undefined;
    to?: string | undefined;
    date?: Date;
    paymentTerms?: string | undefined;
    dueDate?: Date | undefined;
    discount?: number;
    discountType?: DiscountType;
    tax?: number;
    taxType?: TaxType;
    amountPaid?: number;
    invoiceItems?: InvoiceItemViewModel[] | undefined;

    constructor(data?: ICreateInvoiceCommand) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.invoiceNumber = _data["invoiceNumber"];
            this.from = _data["from"];
            this.to = _data["to"];
            this.date = _data["date"] ? new Date(_data["date"].toString()) : <any>undefined;
            this.paymentTerms = _data["paymentTerms"];
            this.dueDate = _data["dueDate"] ? new Date(_data["dueDate"].toString()) : <any>undefined;
            this.discount = _data["discount"];
            this.discountType = _data["discountType"];
            this.tax = _data["tax"];
            this.taxType = _data["taxType"];
            this.amountPaid = _data["amountPaid"];
            if (Array.isArray(_data["invoiceItems"])) {
                this.invoiceItems = [] as any;
                for (let item of _data["invoiceItems"])
                    this.invoiceItems!.push(InvoiceItemViewModel.fromJS(item));
            }
        }
    }

    static fromJS(data: any): CreateInvoiceCommand {
        data = typeof data === 'object' ? data : {};
        let result = new CreateInvoiceCommand();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["invoiceNumber"] = this.invoiceNumber;
        data["from"] = this.from;
        data["to"] = this.to;
        data["date"] = this.date ? this.date.toISOString() : <any>undefined;
        data["paymentTerms"] = this.paymentTerms;
        data["dueDate"] = this.dueDate ? this.dueDate.toISOString() : <any>undefined;
        data["discount"] = this.discount;
        data["discountType"] = this.discountType;
        data["tax"] = this.tax;
        data["taxType"] = this.taxType;
        data["amountPaid"] = this.amountPaid;
        if (Array.isArray(this.invoiceItems)) {
            data["invoiceItems"] = [];
            for (let item of this.invoiceItems)
                data["invoiceItems"].push(item.toJSON());
        }
        return data; 
    }
}

export interface ICreateInvoiceCommand {
    invoiceNumber?: string | undefined;
    from?: string | undefined;
    to?: string | undefined;
    date?: Date;
    paymentTerms?: string | undefined;
    dueDate?: Date | undefined;
    discount?: number;
    discountType?: DiscountType;
    tax?: number;
    taxType?: TaxType;
    amountPaid?: number;
    invoiceItems?: InvoiceItemViewModel[] | undefined;
}

export enum DiscountType {
    Flat = 0,
    Percentage = 1,
}

export enum TaxType {
    Flat = 0,
    Percentage = 1,
}

export class InvoiceItemViewModel implements IInvoiceItemViewModel {
    id?: number;
    item?: string | undefined;
    qty?: number;
    rate?: number;
    amount?: number;

    constructor(data?: IInvoiceItemViewModel) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.item = _data["item"];
            this.qty = _data["qty"];
            this.rate = _data["rate"];
            this.amount = _data["amount"];
        }
    }

    static fromJS(data: any): InvoiceItemViewModel {
        data = typeof data === 'object' ? data : {};
        let result = new InvoiceItemViewModel();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["item"] = this.item;
        data["qty"] = this.qty;
        data["rate"] = this.rate;
        data["amount"] = this.amount;
        return data; 
    }
}

export interface IInvoiceItemViewModel {
    id?: number;
    item?: string | undefined;
    qty?: number;
    rate?: number;
    amount?: number;
}